requirements.txt
numpy
matplotlib
scipy
seaborn
python-dotenv